% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/twister.R
\name{auto_twister}
\alias{auto_twister}
\title{Function for running a game of Twister.}
\usage{
auto_twister(.rounds = 10, .wait = 10, .players = NA, .speech = FALSE)
}
\arguments{
\item{.rounds}{An integer value of the number of rounds to be played.}

\item{.wait}{An integer value of the seconds to wait between play instructions.}

\item{.players}{A character vector containing the names of all players.}

\item{.speech}{A logical value indicating whether the instruction should be
    read aloud by the computer, currently only working under Mac OS.}
}
\value{
A sequence of console messages containing the play instruction and
    accompanying plots that illustrate them.
}
\description{
\code{auto_twister} lets you run a game of Twister with a set of players
for a specified number of rounds.
}
\examples{
auto_twister(.rounds = 2, .wait = 0, .players = c("Alice", "Bob"))
}

