% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/twister.R
\name{twister}
\alias{twister}
\title{Function for spinning the Twister.}
\usage{
twister(.player = NA, .speech = FALSE)
}
\arguments{
\item{.player}{An optional string containing the name of the player.}

\item{.speech}{A logical value indicating whether the instruction should be
    read aloud by the computer, currently only working under Mac OS.}
}
\value{
A console message containing the play instruction and a plot illustrating it.
}
\description{
\code{twister} is the virtual equivalent of running the Twister spinner
and finding out which limb to place on which colour.
}
\examples{
twister("Dude")
}

