% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/twister.R
\name{twister_speech}
\alias{twister_speech}
\title{Helper function for speech output.}
\usage{
twister_speech(.message)
}
\arguments{
\item{.message}{A string containing the play instruction.}
}
\description{
\code{twister_speech} is a helper function that is called in case the play instruction
should be read aloud by the computer.
}

